deleted test_bot, added more training sentences
