+++ src/app/parent_bot.py
@@ -13,10 +13,9 @@ from besser.bot.nlp.intent_classifier.intent_classifier_configuration import LLM
 from besser.bot.nlp.llm.llm_openai_api import LLMOpenAI
 from besser.bot.platforms.websocket import WEBSOCKET_PORT
 
-# Configure the logging module
 logging.basicConfig(level=logging.INFO, format='{levelname} - {asctime}: {message}', style='{')
 
-parent_bot = Bot('parent_bot')
+parent_bot = Bot('llm_bot')
 parent_bot.load_properties('config.ini')
 websocket_platform = parent_bot.use_websocket_platform(use_ui=False)
 
@@ -28,156 +27,253 @@ gpt = LLMOpenAI(
     num_previous_messages=10
 )
 
-
 ic_config = LLMIntentClassifierConfiguration(
     llm_name='gpt-4o-mini',
-    parameters={
-        "seed": None,
-        "top_p": 1,
-        "temperature": 1,
-    },
+    parameters={},
     use_intent_descriptions=True,
-    use_training_sentences=True,
+    use_training_sentences=False,
     use_entity_descriptions=True,
     use_entity_synonyms=False
 )
-
 parent_bot.set_default_ic_config(ic_config)
 
+############################################################################################################################################
+
+# OTHER FUNCTIONS
+
+global user_knows_about_tags
+user_knows_about_tags = False
+
+def updateTags(): #get all the existing tags from datalux
+    url = "https://data.public.lu/api/1/datasets/?format=csv"
+    tags = set()
+    while(url):
+        response = requests.get(url).json()
+        for dataset in response['data']:
+            tags.update(dataset['tags'])
+        url = response['next_page']
+    with open("src/app/datasets_tags.json", "w") as file:
+        dict_tags = {"tags": list(tags)}
+        json.dump(dict_tags, file, indent=4)
+
 # STATES
-###################################################################################################
-initial_state = parent_bot.new_state('initial_parent_botstate', initial=True)
+
+greetings_state = parent_bot.new_state('greetings_state', initial=True)
+help_state = parent_bot.new_state('help_state')
+transition_state = parent_bot.new_state('transition_state')
 smalltalk_state = parent_bot.new_state('smalltalk_state')
-confused_state = parent_bot.new_state('confused_state')
-id_state = parent_bot.new_state('id_state')
-update_state = parent_bot.new_state('update_state')
-#######################################################################################################
+databaseRequest_state = parent_bot.new_state('databaseRequest_state')
+updateTags_state = parent_bot.new_state('updateTags_state')
+answer_state = parent_bot.new_state('answer_state')
 
 # ENTITIES
-#########################################################################################################
-with open("src/app/datasets_id_names.json", "r") as file:
-    name_id_dict = json.load(file)
-id_entries = {key: [] for key in name_id_dict.keys()}
-
-id_entity = parent_bot.new_entity(
-    name='id_entity',
-    description='The ID of a dataset, a long string containing numbers and a few alphabetical characters',
-    entries=id_entries
-)
-
-"""name_entity = parent_bot.new_entity(
-    name='name_entity',
-    description='The name of a dataset, a string of alphabetical characters',
-    entries={
-    'name1': [],
-    'name2': [],
-    }
-)"""
 
 word_entity = parent_bot.new_entity(
     name='word_entity',
-    description='a word, a string of alphabetical characters with meaning',
+    description='a string of letters with meaning. This entity does not include nonsensical string of letters',
 )
-############################################################################################################
 
 # INTENTS
-#############################################################################################################
-id_intent = parent_bot.new_intent('id_intent', [
-    'What is the dataset of id ID?',
-    'id dataset ID',
-    'dataset id ID',
-])
-id_intent.parameter('id1', 'ID', id_entity)
 
-update_intent = parent_bot.new_intent('update_intent', [
-    'update',
+hello_intent = parent_bot.new_intent(
+    name='hello_intent',
+    description='The user greets you'
+)
+
+help_intent = parent_bot.new_intent(
+    name='help_intent',
+    description='The user is asking for help concerning your functionnalities.',
+    training_sentences = [
+    'what can you do ?',
+    'help',
+    'can you help me',
+    'how do I get a specific dataset ?',
+    'how do I ask you for a dataset ?',
+    'What kind of data can you provide ?',
+    'What kind of dataset can you fetch',
+    'What can I ask for',
+    'What is your purpose',
+    'Where are your data coming from',
+    'tell me about your functionnalities',
+    "give me your sources",
+    "Can I get some assistance?",
+    "I need help.",
+    "Help me out here.",
+    "What do you offer?",
+    "How do I use you?",
+    "Explain your features.",
+    "What can I ask you?",
+    "Show me how to get a dataset.",
 ])
 
-database_intent = parent_bot.new_intent('database_intent', [
-    'What is database talks about TOPIC',
-    'id dataset ID',
-    'dataset id ID',
-])
-id_intent.parameter('topic1', 'TOPIC', word_entity)
+databaseRequest_intent = parent_bot.new_intent(
+    name = 'databaseRequest_intent',
+    description = 'The user asks for a dataset about a specific topic',
+    training_sentences = [
+    'What database talks about TOPIC',
+    'Can you give me a TOPIC dataset',
+    'I would like to see datas on TOPIC',
+    'data on TOPIC',
+    "Can you find a dataset about TOPIC?",
+    "I need data on TOPIC.",
+    "Do you have datasets about TOPIC?",
+    "Show me the latest TOPIC data.",
+    "Get me datasets related to TOPIC.",
+    "Find me information about TOPIC.",
+    "I want to see datasets on TOPIC.",
+    "Give me TOPIC data.",
+    "do you have any info about TOPIC?",
+    "Can you pull up data for TOPIC?",
+    "Search for datasets about TOPIC.",
+    ])
+databaseRequest_intent.parameter('topic1', 'TOPIC', word_entity)
 
 smalltalk_intent = parent_bot.new_intent(
     name='smalltalk_intent',
     description='The user is trying to talk about random things that are not related to datasets creation, which are out of your purpose. The questions can be extremely varied.',
     training_sentences = ['how are you doing?',
-     'do you want to chat?'
-     'what is your favorite color?'
-     'give me a cooking recipe'
-     'help me solve this math problem'
-     'is Italy a democraty'
+    'do you want to chat?',
+    'what is your favorite color?',
+    'give me a cooking recipe',
+    'help me solve this math problem',
+    'is Italy a democraty',
+    "How are you today?",
+    "Do you like coffee?",
+    "What's your favorite book?",
+    "What's the weather like?",
+    "Do you play video games?",
+    "Can you tell me a joke?",
+    "Tell me something interesting.",
+    "What do you think of the latest movie?",
+    "Where would you like to travel?",
 ])
-############################################################################################################
 
-# OTHER FUNCTIONS
-
-def updateDatabase():
-    dict = {} #dictionary to link name and id of every dataset
-    url = "https://data.public.lu/api/1/datasets/"
-    while(url):
-        response = requests.get(url).json()
-        for dataset in response['data']:
-            dict[dataset['id']] = dataset['title'] #the ids are unique so they are the keys
-        url = response['next_page']
-    with open("src/app/datasets_id_names.json", "w") as file:
-        json.dump(dict, file, indent=4)
+updateTags_intent = parent_bot.new_intent(
+    name='updateTags_intent',
+    description='The user is asking you to update the list of available tags because they think it might be out of date.',
+    training_sentences = ['Can you update the tags?',
+    'Update the tags.',
+    'Keep your list of available tags up to date.',
+    'Get the new tags.',
+    'You should update your tags.',
+    "Can you refresh the tags list?",
+    "Please update your tags.",
+    "Are the tags up-to-date?",
+    "I think the tags need updating.",
+    "Can you get the latest tags?",
+    "Do you have the newest tags?",
+    "Please sync the tags.",
+    "Can you update the dataset tags?",
+    "Get the most recent tags.",
+    "Please update the available tags list.",
+])
 
 
 # STATES BODIES' DEFINITION + TRANSITIONS
 
-def confused_body(session: Session):
-    #answer = gpt.predict(f"You are being used within an intent-based chatbot. Your goal is to help the user browse data websites to find relevant datasets for their needs. You are generating a fallback answer because the intent of the user was not identified. Generate a message similar to 'I am sorry, I did not quite catch what you were trying to ask here. Can you rephrase it ?', based on the user message: {session.message}")
-    #session.reply(answer)
-    session.reply("I didn't understand what you are trying to ask.")
-    
+def global_fallback_body(session: Session):
+    answer = gpt.predict(f"You are being used within an intent-based bot. The chatbot triggered the fallback mechanism because no intent was recognized from the user input. Generate a message similar to 'Sorry, I don't know the answer', based on the user message: {session.message}")
+    session.reply(answer)
+
+
+parent_bot.set_global_fallback_body(global_fallback_body)
+
+
+def greetings_body(session: Session):
+    answer = gpt.predict(f"You are being used within an intent-based chatbot. Your goal is to help the user browse data websites to find relevant datasets for their needs. Start the conversation with a short (5-20 words) greetings message asking the user what they need.")
+    session.reply(answer)
+
+
+greetings_state.set_body(greetings_body)
+greetings_state.when_intent_matched_go_to(hello_intent, greetings_state)
+greetings_state.when_intent_matched_go_to(smalltalk_intent, smalltalk_state)
+greetings_state.when_intent_matched_go_to(databaseRequest_intent, databaseRequest_state)
+greetings_state.when_intent_matched_go_to(updateTags_intent, updateTags_state)
+
+def help_body(session: Session):
+    answer = gpt.predict(
+        f"You are being used within an intent-based chatbot. The user is asking for help concerning your functionnalities. Give an answer based on the user message : {session.message}, and using the following informations :"
+        f"You are an helpful assistant chatbot, that has access to the datasets of the luxembourgish open data platform (data.public.lu). You can provide the user with a dataset suited to their needs if they give you keywords on what they are looking for."
+        f"You can then generate another chatbot trained on the dataset the user chose, and who will be able to answer any question concerning this specific dataset."
+        )
+    session.reply(answer)
+
+
+help_state.set_body(help_body)
+help_state.set_global(help_intent)
+help_state.go_to(transition_state)
+
+def transition_body(session: Session):
+    session.reply("Do you need anything else ?")
+
+
+transition_state.set_body(transition_body)
+transition_state.when_intent_matched_go_to(smalltalk_intent, smalltalk_state)
+transition_state.when_intent_matched_go_to(databaseRequest_intent, databaseRequest_state)
+transition_state.when_intent_matched_go_to(updateTags_intent, updateTags_state)
 
 def smalltalk_body(session: Session):
     answer = gpt.predict(f"You are being used within an intent-based chatbot. Your goal is to help the user browse data websites to find relevant datasets for their needs. This answer is generated because the user attempted to make small talk or to ask a question that has nothing to do with the aim of the chatbot. Generate a message similar to 'That seems very interesting, but my purpose is to help you find datasets relevant to your needs, not this.', based on the user message: {session.message}")
     session.reply(answer)
 
-smalltalk_state.set_global(smalltalk_intent)
-smalltalk_state.when_intent_matched_go_to(id_intent, id_state)
+smalltalk_state.set_body(smalltalk_body)
+smalltalk_state.go_to(transition_state)
 
-def initial_body(session: Session):
-    session.reply("Hello! Give me a dataset ID or name and I will retrieve it!")
-
-initial_state.set_body(initial_body)
-initial_state.when_intent_matched_go_to(id_intent, id_state)
-initial_state.when_intent_matched_go_to(update_intent, update_state)
-parent_bot.set_global_fallback_body(confused_body)
-
-
-def id_body(session: Session):
+def databaseRequest_body(session: Session):
+    global user_knows_about_tags
     predicted_intent: IntentClassifierPrediction = session.predicted_intent
-    id = predicted_intent.get_parameter('id1')
-    if id.value is None:
-        session.reply("Sorry, it seems this id does not exist. If you think this is a mistake you can ask me to update my database.")
+    topic = predicted_intent.get_parameter('topic1')
+    if topic.value is None:
+        session.reply("Sorry, it seems this tag doesn't mean anything.")
     else:
-        url = "https://data.public.lu/api/1/datasets/" + str(id.value) + "/"
+        url = "https://data.public.lu/api/1/datasets/?tag=" + str(topic.value) + "&format=csv"
         response = requests.get(url).json()
-        format = ""
-        for resource in response['resources']:
-                if resource['format'] == "csv":
-                    format = resource['format']
-                    break
-        if format == "csv":
-            session.reply(f"It seems the {id.value} dataset has a csv. I will generate infos on the right side of the screen.")
-            session.reply(f"{resource['url']}")
-        else:
-            session.reply(f"It seems the {id.value} dataset does not have any csv.")
+        if not response['data']:
+            with open("src/app/datasets_tags.json", 'r') as file:
+                available_tags = json.load(file)
+            tags_set = set(available_tags["tags"])
+            tags_str = ", ".join(str(tag) for tag in tags_set)
+            if user_knows_about_tags == False:
+                answer = gpt.predict(
+                    f"You are being used within an intent-based chatbot. The user asked you to provide a dataset with this specific tag '{topic.value}' but you found none. "
+                    f"Here is the list of the only available tags: {tags_str}. Give them a possible synonym to their first demand if one is in the list."
+                    f"Tell them that they can ask you to update your tag list if they think it is not up to date."
+                    )
+                user_knows_about_tags = True
+            else:
+                answer = gpt.predict(
+                    f"You are being used within an intent-based chatbot. The user asked you to provide a dataset with this specific tag '{topic.value}' but you found none. "
+                    f"Here is the list of the only available tags: {tags_str}. Give them a possible synonym to their first demand if one is in the list."
+                    )
+            session.reply(answer)
+            return
+        else :
 
-id_state.set_body(id_body)
-id_state.set_global(id_state)
-id_state.go_to(initial_state)
 
-def update_body(session: Session):
-    session.reply('I will start updating my database and tell you when I am done. Please wait, it will take a little while.')
-    updateDatabase()
-    session.reply('The update is done !')
+            session.reply(f"I found {len(response['data'])} result(s) mentioning {topic.value}:")
+            all_datasets_info = []
+            for dataset in response['data']:
+                for resource in dataset['resources']:
+                    if resource['format'] == "csv":
+                        useful_info_dict = {
+                            "dataset_title": resource["title"],
+                            "dataset_date": resource["created_at"],
+                            "dataset_description": resource["description"],
+                            "dataset_url": resource["url"]
+                        }
+                        all_datasets_info.append(useful_info_dict)
+                        break
+            all_datasets_info_json = json.dumps(all_datasets_info, indent=4)
+            session.reply(all_datasets_info_json)
+            
+databaseRequest_state.set_body(databaseRequest_body)
+databaseRequest_state.go_to(transition_state)
 
-update_state.set_body(update_body)
-update_state.set_global(update_state)
-update_state.go_to(initial_state)
+def updateTags_body(session: Session):
+    session.reply(f"I am updating my tag list. Please wait, it might take a while. I will tell you when I am done.")
+    updateTags()
+    session.reply(f"The update is over!")
+
+
+updateTags_state.set_body(updateTags_body)
+updateTags_state.go_to(transition_state)
